Reorganize application mode logic
