+++ app/controllers/application_controller.rb
@@ -7,6 +7,8 @@ class ApplicationController < ActionController::Base
   before_action :authenticate_user!, :check_bundle_installed, :check_backend_jobs, except: [:page_not_found, :server_error]
   around_action :catch_not_found
 
+  helper_method :mode_internal?, :mode_demo?, :mode_atl?
+
   rescue_from CanCan::AccessDenied do |exception|
     render text: exception, status: 401
   end
@@ -23,6 +25,25 @@ class ApplicationController < ActionController::Base
       format.html { render template: 'errors/500', layout: 'layouts/errors', status: 500 }
       format.all  { render text: '500 Server Error', status: 500 }
     end
+  end
+
+  def mode_internal?
+    Settings[:auto_approve] && Settings[:ignore_roles] && Settings[:enable_debug_features] && Settings[:default_role].nil?
+  end
+
+  def mode_demo?
+    Settings[:auto_approve] && !Settings[:ignore_roles] && Settings[:enable_debug_features] && Settings[:default_role] == :user
+  end
+
+  def mode_atl?
+    !Settings[:auto_approve] && !Settings[:ignore_roles] && !Settings[:enable_debug_features] && Settings[:default_role].nil?
+  end
+
+  def application_mode
+    return 'Internal' if mode_internal?
+    return 'Demo' if mode_demo?
+    return 'ATL' if mode_atl?
+    'Custom'
   end
 
   private
